Combined block service and block propose. Edited test case.
